vmm: Allow for allocating specific PCI BDFs
We replace `next_device_bdf` with `allocate_device_bdf`. The new
function optionally accepts a device ID to use when allocating a BDF.
If not provided, we fall back to allocating the next available BDF.
This allows for creating specific BDFs on the respective segment.

Signed-off-by: Pascal Scholz <pascal.scholz@cyberus-technology.de>
On-behalf-of: SAP pascal.scholz@sap.com